requirements.txt
aiohttp~=3.8.4
email-validator==2.0.0.post2
fastapi==0.95.2
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
secure==0.3.0
ujson==5.6.0
uvicorn==0.22.0
gunicorn~=20.1.0
